login: move welcome-test rewards down; empty both ratchet allowlists
The last two pieces of business logic leave endpoints/login.py:

  - compute_score_of_form_response joins its sibling
    services/forms.validate_form_response -- same domain, same field-type
    dispatch, and it is form scoring rather than anything login-specific.
  - the claim flow becomes services/welcome_test.claim_welcome_test_rewards.
    Its own module rather than part of services/rewards.py, which owns the
    answer-question Reward model; this pays a CoinDeposit and touches no
    Reward.

That removes login.py's last db.commit() (was :484, the last statement before
the return, so a no-op on the success path) and its last `crud` import. The
file is now 181 lines of routing: every handler parses the request, resolves
its deps, calls one service function and returns its schema.

With login.py clean, BOTH ratchet allowlists are now empty:

  - check_layer_imports.py: `allow` is an empty set; api/ no longer imports
    crud or responders anywhere.
  - check_service_commits.py: ALLOWLIST is an empty set; api/, services/ and
    crud/ hold zero commits, and the only remaining transaction boundaries are
    api/deps.py and infra/runtime.py.

Both allowlist mechanisms are kept in place (widening the ratchet is step 4,
item 7). check_service_commits already errors on an unused entry, so its
allowlist cannot silently regrow.
